Started CMD - Command Prompt course on YouTube

+++ 100_days_of_code_udemy_dr_angela_yu/section_002/scction_002_project_tip_calculator.py
@@ -6,7 +6,7 @@
 
 
 print("Welcome to the tip calculator!")
-bill = int(input("What was the total bill? $"))
+bill = float(input("What was the total bill? $"))
 tip = int(input("How much tip would you like to give? 10, 12, or 15?"))
 tip_amount = float((tip / bill) * 100)
 people = int(input("How many people to split the bill?"))
